First pass at extracting the Coupon Excel functions
