+++ tests/PhpSpreadsheetTests/Calculation/Functions/Financial/NominalTest.php
@@ -17,10 +17,12 @@ class NominalTest extends TestCase
      * @dataProvider providerNOMINAL
      *
      * @param mixed $expectedResult
+     * @param mixed $rate
+     * @param mixed $periods
      */
-    public function testNOMINAL($expectedResult, ...$args): void
+    public function testNOMINAL($expectedResult, $rate, $periods): void
     {
-        $result = Financial::NOMINAL(...$args);
+        $result = Financial::NOMINAL($rate, $periods);
         self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
     }
